Uses UnobservedTCARouter when Screen is not ObservableState
