+++ Sources/TCACoordinators/Collection+safeSubscript.swift
@@ -1,0 +1,6 @@
+extension Collection {
+  /// Returns the element at the specified index if it is within bounds, otherwise nil.
+  subscript(safe index: Index) -> Element? {
+    indices.contains(index) ? self[index] : nil
+  }
+}
