+++ Sources/TCACoordinators/TCARouter/TCARouter+IndexedScreen.swift
@@ -16,17 +16,3 @@ public extension TCARouter where ID == Int {
     )
   }
 }
-
-public extension ObservedTCARouter where ID == Int {
-  /// Convenience initializer for managing screens in an `Array`, identified by index.
-  init(
-    _ store: Store<[Route<Screen>], IndexedRouterAction<Screen, ScreenAction>>,
-    @ViewBuilder screenContent: @escaping (Store<Screen, ScreenAction>) -> ScreenContent
-  ) {
-    self.init(
-      store: store,
-      identifier: { $1 },
-      screenContent: screenContent
-    )
-  }
-}
