+++ TCACoordinatorsExample/TCACoordinatorsExample/Game/LogInCoordinator.swift
@@ -12,7 +12,7 @@ struct LogInCoordinatorView: View {
   let store: StoreOf<LogInCoordinator>
 
   var body: some View {
-    ObservedTCARouter(store.scope(state: \.routes, action: \.router)) { screen in
+    TCARouter(store.scope(state: \.routes, action: \.router)) { screen in
       switch screen.case {
       case let .welcome(store):
         WelcomeView(store: store)
